refactor: move database code to `Importer` project
