+++ Importer.CLI/Importer.CLI.csproj
@@ -4,10 +4,6 @@
     <ProjectReference Include="..\Importer\Importer.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Data.Sqlite" Version="6.0.6" />
-  </ItemGroup>
-
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
