+++ Importer/DatabaseWriter.cs
@@ -1,4 +1,4 @@
-namespace Importer.CLI;
+namespace Importer;
 
 using Microsoft.Data.Sqlite;
