+++ Importer/Importer.csproj
@@ -10,4 +10,8 @@
     <InternalsVisibleTo Include="Importer.Test" />
   </ItemGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Data.Sqlite" Version="6.0.6" />
+  </ItemGroup>
+
 </Project>
